Add Javadoc and move MovementAPI singleton instance to it's own class
