+++ core/src/main/java/nl/tudelft/contextproject/core/config/Constants.java
@@ -1,5 +1,9 @@
 package nl.tudelft.contextproject.core.config;
 
+/**
+ * This class serves as an internal "configuration" for parameters that might be useful in
+ * multiple (independent) locations or parameters that are likely to become subject to change.
+ */
 public abstract class Constants {
     // The Width and Height of the camera view
     // This corresponds with the x and y axes of the screen
