+++ desktop/src/main/java/nl/tudelft/contextproject/java/MainDesktop.java
@@ -6,6 +6,11 @@ import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import nl.tudelft.contextproject.core.Main;
 
 public class MainDesktop {
+
+    /**
+     * Main entry point of the desktop client.
+     * @param args The program arguments
+     */
     public static void main(String[] args) {
         LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
         config.title = "Human Pencils";
